Struct vp8::parser::Header

pub struct Header {

    key_frame: bool,
    version: u8,
    show_frame: bool,
    data_chunk_size: u8,
    first_part_size: u32,
    width: u16,
    height: u16,
    horiz_scale_code: u8,
    vert_scale_code: u8,
    color_space: bool,
    clamping_type: bool,
    filter_type: bool,
    loop_filter_level: u8,
    sharpness_level: u8,
    log2_nbr_of_dct_partitions: u8,
    partition_size: [u32; 8],
    quant_indices: QuantIndices,
    refresh_entropy_probs: bool,
    refresh_last: bool,
    refresh_golden_frame: bool,
    refresh_alternate_frame: bool,
    copy_buffer_to_golden: u8,
    copy_buffer_to_alternate: u8,
    sign_bias_golden: bool,
    sign_bias_alternate: bool,
    coeff_prob: [[[[u8; 11]; 3]; 8]; 4],
    mv_prob: [[u8; 19]; 2],
    mb_no_coeff_skip: bool,
    prob_skip_false: u8,
    prob_intra: u8,
    prob_last: u8,
    prob_golden: u8,
    mode_probs: ModeProbs,
    bd_range: usize,
    bd_value: usize,
    bd_count: isize,
    header_size: u32,

}

Fields
key_frame: bool

Indicates if the current frame is a key frame or not.

version: u8

Determines the bitstream version.

show_frame: bool

Indicates if the current frame is meant to be displayed or not.

data_chunk_size: u8

The size in bytes of the Uncompressed Data Chunk

first_part_size: u32

Determines the size of the first partition (control partition) excluding the size of the Uncompressed Data Chunk

width: u16

The frame’s width, in pixels.

height: u16

The frame’s height, in pixels.

horiz_scale_code: u8

Horizontal scale code value.

vert_scale_code: u8

Vertical scale code value.

color_space: bool

Defines the YUV color space of the sequence.

clamping_type: bool

Specifies if the decoder is required to clamp the reconstructed pixel values.

filter_type: bool

Determines whether the normal or the simple loop filter is used.

loop_filter_level: u8

Controls the deblocking filter.

sharpness_level: u8

Controls the deblocking filter.

log2_nbr_of_dct_partitions: u8

Determines the number of separate partitions containing the DCT coefficients of the macroblocks.

partition_size: [u32; 8]quant_indices: QuantIndices

Dequantizer indices.

refresh_entropy_probs: bool

Determines whether updated token probabilities are used only for this frame or until further update

refresh_last: bool

Determines if the current decoded frame refreshes the last frame reference buffer

refresh_golden_frame: bool

Determines if the current decoded frame refreshes the golden frame.

refresh_alternate_frame: bool

Determines if the current decoded frame refreshes the alternate reference frame.

copy_buffer_to_golden: u8

Determines if the golden reference is replaced by another reference.

copy_buffer_to_alternate: u8

Determines if the alternate reference is replaced by another reference.

sign_bias_golden: bool

Controls the sign of motion vectors when the golden frame is referenced.

sign_bias_alternate: bool

Controls the sign of motion vectors when the alternate frame is referenced.

coeff_prob: [[[[u8; 11]; 3]; 8]; 4]

The new branch probability for the DCT/WHT tree.

mv_prob: [[u8; 19]; 2]

MV decoding probability.

mb_no_coeff_skip: bool

Enables or disables the skipping of macroblocks containing no non-zero coefficients.

prob_skip_false: u8

The probability that the macroblock is not skipped (flag indicating skipped macroblock is false).

prob_intra: u8

The probability of an intra macroblock.

prob_last: u8

The probability that the last reference frame is used for inter prediction.

prob_golden: u8

The probability that the golden reference frame is used for inter prediction.

mode_probs: ModeProbs

Branch probabilities kept live across frames.

bd_range: usize

Boolean decoder range for this frame.

bd_value: usize

Boolean decoder value for this frame.

bd_count: isize

Boolean decoder count for this frame.

header_size: u32

The size in bits of the Frame Header, thus excluding any Uncompressed Data Chunk bytes.

Implementations

impl Header

pub fn key_frame(&self) -> bool

Get a reference to the header’s key frame.

pub fn version(&self) -> u8

Get a reference to the header’s version.

pub fn show_frame(&self) -> bool

Get a reference to the header’s show frame.

pub fn data_chunk_size(&self) -> u8

Get a reference to the header’s data chunk size.

pub fn first_part_size(&self) -> u32

Get a reference to the header’s first part size.

pub fn width(&self) -> u16

Get a reference to the header’s width.

pub fn height(&self) -> u16

Get a reference to the header’s height.

pub fn horiz_scale_code(&self) -> u8

Get a reference to the header’s horiz scale code.

pub fn vert_scale_code(&self) -> u8

Get a reference to the header’s vert scale code.

pub fn color_space(&self) -> bool

Get a reference to the header’s color space.

pub fn clamping_type(&self) -> bool

Get a reference to the header’s clamping type.

pub fn filter_type(&self) -> bool

Get a reference to the header’s filter type.

pub fn loop_filter_level(&self) -> u8

Get a reference to the header’s loop filter level.

pub fn sharpness_level(&self) -> u8

Get a reference to the header’s sharpness level.

pub fn log2_nbr_of_dct_partitions(&self) -> u8

Get a reference to the header’s log2 nbr of dct partitions.

pub fn partition_size(&self) -> [u32; 8]

Get a reference to the header’s partition size.

pub fn quant_indices(&self) -> &QuantIndices

Get a reference to the header’s quant indices.

pub fn refresh_entropy_probs(&self) -> bool

Get a reference to the header’s refresh entropy probs.

pub fn refresh_last(&self) -> bool

Get a reference to the header’s refresh last.

pub fn refresh_golden_frame(&self) -> bool

Get a reference to the header’s refresh golden frame.

pub fn refresh_alternate_frame(&self) -> bool

Get a reference to the header’s refresh alternate frame.

pub fn copy_buffer_to_golden(&self) -> u8

Get a reference to the header’s copy buffer to golden.

pub fn copy_buffer_to_alternate(&self) -> u8

Get a reference to the header’s copy buffer to alternate.

pub fn sign_bias_golden(&self) -> bool

Get a reference to the header’s sign bias golden.

pub fn sign_bias_alternate(&self) -> bool

Get a reference to the header’s sign bias alternate.

pub fn coeff_prob(&self) -> [[[[u8; 11]; 3]; 8]; 4]

Get a reference to the header’s coeff prob.

pub fn mv_prob(&self) -> [[u8; 19]; 2]

Get a reference to the header’s mv prob.

pub fn mb_no_coeff_skip(&self) -> bool

Get a reference to the header’s mb no coeff skip.

pub fn prob_skip_false(&self) -> u8

Get a reference to the header’s prob skip false.

pub fn prob_intra(&self) -> u8

Get a reference to the header’s prob intra.

pub fn prob_last(&self) -> u8

Get a reference to the header’s prob last.

pub fn prob_golden(&self) -> u8

Get a reference to the header’s prob golden.

pub fn mode_probs(&self) -> ModeProbs

Get a reference to the header’s mode probs.

pub fn bd_range(&self) -> usize

Get a reference to the header’s bd range.

pub fn bd_value(&self) -> usize

Get a reference to the header’s bd value.

pub fn bd_count(&self) -> isize

Get a reference to the header’s bd count.

pub fn header_size(&self) -> u32

Get a reference to the header’s header size.